+++ pypdf/_page.py
@@ -92,7 +92,7 @@ MERGE_CROP_BOX = "cropbox"  # pypdf <= 3.4.0 used "trimbox"
 
 
 def _get_rectangle(self: Any, name: str, defaults: Iterable[str]) -> RectangleObject:
-    retval: Union[None, RectangleObject, IndirectObject] = self.get(name)
+    retval: Union[None, RectangleObject, ArrayObject, IndirectObject] = self.get(name)
     if isinstance(retval, RectangleObject):
         return retval
     if is_null_or_none(retval):
@@ -102,7 +102,11 @@ def _get_rectangle(self: Any, name: str, defaults: Iterable[str]) -> RectangleOb
                 break
     if isinstance(retval, IndirectObject):
         retval = self.pdf.get_object(retval)
-    retval = RectangleObject(retval)  # type: ignore
+    if isinstance(retval, ArrayObject) and (length := len(retval)) > 4:
+        logger_warning(f"Expected four values, got {length}: {retval}", __name__)
+        retval = RectangleObject(tuple(retval[:4]))
+    else:
+        retval = RectangleObject(retval)  # type: ignore
     _set_rectangle(self, name, retval)
     return retval
 
@@ -374,10 +378,8 @@ class ImageFile:
             )
 
         from ._reader import PdfReader  # noqa: PLC0415
-
-        # to prevent circular import
-        from ._xobj_image_helpers import _xobj_to_image  # noqa: PLC0415
         from .generic import DictionaryObject, PdfObject  # noqa: PLC0415
+        from .generic._image_xobject import _xobj_to_image  # noqa: PLC0415
 
         if self.indirect_reference is None:
             raise TypeError("Cannot update an inline image.")
@@ -650,7 +652,7 @@ class PageObject(DictionaryObject):
                     raise KeyError("No inline image can be found")
                 return self.inline_images[id]
 
-            from ._xobj_image_helpers import _xobj_to_image  # noqa: PLC0415
+            from .generic._image_xobject import _xobj_to_image  # noqa: PLC0415
             imgd = _xobj_to_image(cast(DictionaryObject, xobjs[id]))
             extension, byte_stream = imgd[:2]
             return ImageFile(
@@ -754,7 +756,7 @@ class PageObject(DictionaryObject):
                 if k not in init:
                     init[k] = v
             ii["object"] = EncodedStreamObject.initialize_from_dictionary(init)
-            from ._xobj_image_helpers import _xobj_to_image  # noqa: PLC0415
+            from .generic._image_xobject import _xobj_to_image  # noqa: PLC0415
             extension, byte_stream, img = _xobj_to_image(ii["object"])
             files[f"~{num}~"] = ImageFile(
                 name=f"~{num}~{extension}",